backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
asyncpg
alembic[asyncio]
python-dotenv
psycopg2-binary>=2.9
python-dotenv
